# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/gatherings": {
      "get": {
        "description": "장르·세션 필터, 페이지네이션",
        "operationId": "JammitGatheringController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "0부터",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "정렬 키 (백엔드 규약)",
            "schema": {}
          },
          {
            "name": "sessions",
            "required": false,
            "in": "query",
            "description": "필터 세션 (복수 가능)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "VOCAL",
                  "ELECTRIC_GUITAR",
                  "DRUM",
                  "ACOUSTIC_GUITAR",
                  "BASS",
                  "STRING_INSTRUMENT",
                  "PERCUSSION",
                  "KEYBOARD"
                ]
              }
            }
          },
          {
            "name": "genres",
            "required": false,
            "in": "query",
            "description": "필터 장르 (복수 가능)",
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "ROCK",
                  "METAL",
                  "POP",
                  "BALLAD",
                  "INDIE",
                  "ALTERNATIVE",
                  "JAZZ",
                  "PUNK",
                  "ACOUSTIC",
                  "FOLK",
                  "RNB"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "모임 목록 페이지"
          }
        },
        "summary": "모임 목록(공개)",
        "tags": [
          "Jammit Gathering"
        ]
      },
      "post": {
        "operationId": "JammitGatheringController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "주말 합주",
                  "thumbnail": "https://example.com/thumb.jpg",
                  "place": "서울",
                  "description": "초보도 환영",
                  "gatheringDateTime": "2026-05-01T14:00:00.000Z",
                  "recruitDateTime": "2026-04-25T23:59:59.000Z",
                  "genres": [
                    "ROCK",
                    "POP"
                  ],
                  "gatheringSessions": [
                    {
                      "bandSession": "VOCAL",
                      "recruitCount": 2
                    },
                    {
                      "bandSession": "DRUM",
                      "recruitCount": 1
                    }
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "생성된 모임 (CommonResponseDto)"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "모임 생성",
        "tags": [
          "Jammit Gathering"
        ]
      }
    },
    "/gatherings/my/created": {
      "get": {
        "operationId": "JammitGatheringController_myCreated",
        "parameters": [
          {
            "name": "includeCanceled",
            "required": false,
            "in": "query",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내가 만든 모임 목록",
        "tags": [
          "Jammit Gathering"
        ]
      }
    },
    "/gatherings/my/participations": {
      "get": {
        "operationId": "JammitGatheringController_myParticipations",
        "parameters": [
          {
            "name": "includeCanceled",
            "required": false,
            "in": "query",
            "schema": {
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "size",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내가 참가 신청·참여한 모임 목록",
        "tags": [
          "Jammit Gathering"
        ]
      }
    },
    "/gatherings/my/completed": {
      "get": {
        "operationId": "JammitGatheringController_myCompletedGatherings",
        "parameters": [],
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "내가 완료한 모임 목록 (영상 업로드 연동 등)",
        "tags": [
          "Jammit Gathering"
        ]
      }
    },
    "/gatherings/{id}": {
      "get": {
        "operationId": "JammitGatheringController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "gathering id",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "모임 상세"
          },
          "404": {
            "description": "없음"
          }
        },
        "summary": "모임 상세(공개)",
        "tags": [
          "Jammit Gathering"
        ]
      },
      "put": {
        "description": "개설자만 가능",
        "operationId": "JammitGatheringController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "name": "수정된 이름",
                  "place": "부산",
                  "description": "설명 변경"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          },
          "401": {
            "description": "미인증"
          },
          "403": {
            "description": "권한 없음"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "모임 수정",
        "tags": [
          "Jammit Gathering"
        ]
      },
      "delete": {
        "description": "개설자만 가능",
        "operationId": "JammitGatheringController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "모임 취소",
        "tags": [
          "Jammit Gathering"
        ]
      }
    },
    "/gatherings/{id}/complete": {
      "put": {
        "description": "개설자만 가능",
        "operationId": "JammitGatheringController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          },
          "401": {
            "description": "미인증"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "모임 완료 처리",
        "tags": [
          "Jammit Gathering"
        ]
      }
    },
...
    "/review/gathering/{gatheringId}": {
      "get": {
        "operationId": "JammitReviewController_byGathering",
        "parameters": [
          {
            "name": "gatheringId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CommonResponseDto"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "모임별 리뷰 목록",
        "tags": [
          "Jammit Review"
        ]
      }
    },
...
  },
  "info": {
    "title": "jammit api문서",
    "description": "jammit 개발을 위한 API 문서입니다.\n\n🔗 OpenAPI 명세 다운로드: /openapi-spec.json",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "JWT access token",
        "in": "header"
      }
    },
...
  }
}